od: format output like other implementations
In many (most?) cases, the posixutils-rs implementation of `od` is
formatting its output differently from all other implementations I
have ready access to. POSIX does not specify the correct format of
`od` on a character by character level, but in order to maximize the
usefulness of this implementation, it probably makes sense to mirror
the output of other implementations, for the sake of compatibility.

Also: fix a bug in which using the "-j" option does not affect the
input offset that is printed. All other implementations of `od` that
I tested take the skipped bytes into account. However, in one of the
"informative" sections of the POSIX specification, an example seems
to suggest that the input offset should always start at 0, even when
"-j" is used to skip bytes.